requirements.txt
# ======================
# 4. REQUIREMENTS.TXT
# ======================
# Save this as requirements.txt
streamlit==1.28.0
tensorflow==2.12.0
scikit-learn==1.2.2
optuna==3.3.0
opencv-python==4.7.0.72
Pillow==9.5.0
numpy==1.24.3
pandas==1.5.3
seaborn==0.12.2